+++ packages/exporters/src/firebender/index.ts
@@ -12,7 +12,7 @@ import type {
   ResolvedScope,
 } from "../types.js";
 import type { AlignRule } from "@aligntrue/schema";
-import { canonicalizeJson, computeHash } from "@aligntrue/schema";
+import { computeContentHash } from "@aligntrue/schema";
 import { AtomicFileWriter } from "@aligntrue/file-utils";
 import { extractModeConfig, applyRulePrioritization } from "../utils/index.js";
 
@@ -58,8 +58,7 @@ export class FirebenderExporter implements ExporterPlugin {
     );
 
     const allRulesIR = this.state.allRules.map(({ rule }) => rule);
-    const irContent = JSON.stringify({ rules: allRulesIR });
-    const contentHash = computeHash(canonicalizeJson(irContent));
+    const contentHash = computeContentHash({ rules: allRulesIR });
 
     const fidelityNotes = this.computeFidelityNotes(allRulesIR);
 
@@ -127,8 +126,7 @@ export class FirebenderExporter implements ExporterPlugin {
       }));
 
     const allRulesIR = this.state.allRules.map(({ rule }) => rule);
-    const irContent = JSON.stringify({ rules: allRulesIR });
-    const contentHash = computeHash(canonicalizeJson(irContent));
+    const contentHash = computeContentHash({ rules: allRulesIR });
     const fidelityNotes = this.computeFidelityNotes(allRulesIR);
 
     const output = {